+++ R/model_biases.R
@@ -1,6 +1,6 @@
-#' Analysis of future projections
+#' Visualization  of model biases
 #'
-#' Automatically process climate model projections and compute useful statistics
+#' Automatically compute the difference between observations and the historical experiment
 
 #' @param data output of load_data
 #' @param bias.correction logical
@@ -16,8 +16,7 @@
 #' @export
 
 
-
-projections <-
+model_biases <-
   function(data,
            bias.correction = F,
            uppert = NULL,
@@ -38,9 +37,6 @@ projections <-
                duration) {
         stopifnot(is.logical(consecutive), is.logical(bias.correction))
         match.arg(duration, c("max", "total"))
-        if (length(data[[1]]$experiment) == 1 &
-            data[[1]]$experiment[[1]] == "historical")
-          cli::cli_abort(c("x" = "Projections are not part of CAVAanalytics list"))
         if (!is.null(lowert) &
             !is.null(uppert))
           cli::cli_abort(c("x" = "select only one threshold"))
@@ -61,23 +57,19 @@ projections <-
         if (consecutive &
             is.null(uppert) &
             is.null(lowert))
-          stop("Specify a threshold for which you want to calculate consecutive days")
-        if (!any(stringr::str_detect(colnames(data[[1]]), "obs")) &
-            isTRUE(bias.correction)) {
+          cli::cli_abort("Specify a threshold for which you want to calculate consecutive days")
+        if (!any(stringr::str_detect(colnames(data[[1]]), "obs")) |
+            !any(stringr::str_detect(data[[1]]$experiment, "historical"))) {
           cli::cli_abort(
-            c("x" = "Bias correction cannot be performed, no observational dataset found. Set as F")
+            c("x" = "This function requires an observational dataset and the historical experiment to calculate model biases")
           )
 
         }
-        if (bias.correction) {
           if (length(data[[1]]$obs[[1]]$xy$x) != length(data[[1]]$models_mbrs[[1]]$xy$x)) {
             cli::cli_alert_warning(
               "Observation and historical experiment do not have the same spatial resolution. Models will be interpolated to match the observational dataset"
             )
           }
-
-        }
-
       }
 
     # generate messages on the type of operations being performed
@@ -90,7 +82,7 @@ projections <-
                duration) {
         if (is.null(uppert) & is.null(lowert)) {
           paste0(
-            "Calculation of ",
+            "Calculation of model biases for ",
             ifelse(var == "pr", "total ", "mean "),
             ifelse(bias.correction, "bias-corrected ", " "),
             var
@@ -99,7 +91,7 @@ projections <-
         else if ((!is.null(uppert) |
                   !is.null(lowert)) & !consecutive) {
           paste0(
-            "Calculation of number of days with ",
+            "Calculation of model biases for number of days with ",
             var,
             ifelse(
               !is.null(lowert),
@@ -113,7 +105,7 @@ projections <-
                   !is.null(lowert)) &
                  (consecutive & duration == "max")) {
           paste0(
-            "Calculation of maximum length of consecutive number of days ",
+            "Calculation of model biases for maximum length of consecutive number of days ",
             ifelse(
               !is.null(lowert),
               paste0("below ", lowert),
@@ -127,7 +119,7 @@ projections <-
                  (consecutive & duration == "total")) {
           paste0(
             var,
-            ". Calculation of total total number of consecutive days with duration longer than 6 days, ",
+            ". Calculation of model biases for total total number of consecutive days with duration longer than 6 days, ",
             ifelse(
               !is.null(lowert),
               paste0("below threshold of ", lowert),
@@ -147,18 +139,30 @@ projections <-
           "Some data will be lost on year-crossing season subset (see the 'Time slicing' section of subsetGrid documentation for more details)"
         )
       }
-      if (any(stringr::str_detect(colnames(datasets), "obs"))) {
-        datasets %>%  dplyr::mutate_at(c("models_mbrs", "obs"),
-                                       ~ purrr::map(., ~ suppressMessages(
-                                         transformeR::subsetGrid(., season = season)
-                                       )))
+
+      lon <- datasets$models_mbrs[[1]]$xyCoords$x
+      lon_obs <- datasets$obs[[1]]$xyCoords$x
+
+      if (length(lon) != length(lon_obs)) {
+        datasets <- datasets %>%
+          dplyr::filter(experiment == "historical") %>%
+          dplyr::mutate(
+            models_mbrs= map(models_mbrs,~ suppressMessages(transformeR::interpGrid(.x, new.coordinates = transformeR::getGrid(obs[[1]]))))
+          )
       } else {
-        datasets %>%  dplyr::mutate_at(c("models_mbrs"),
-                                       ~ purrr::map(., ~ suppressMessages(
-                                         transformeR::subsetGrid(., season = season)
-                                       )))
+        cli::cli_alert_warning(
+          "Observations and model simulations have the same spatial resolution, proceeding with calculations"
+        )
       }
+      datasets %>%
+        dplyr::filter(experiment == "historical") %>%
+        dplyr::mutate_at(c("models_mbrs", "obs"),
+                         ~ purrr::map(., ~ suppressMessages(
+                           transformeR::subsetGrid(., season = season)
+                         )))
+
     }
+
 
     # function used to perform the calculations
 
@@ -173,11 +177,10 @@ projections <-
                duration,
                country_shp) {
         data_list <- datasets %>%
-          dplyr::filter(experiment != "historical") %>%
           {
             if (bias.correction) {
-              cli::cli_text("{cli::symbol$arrow_right}",
-                paste(
+              cli::cli_text(
+                paste("{cli::symbol$arrow_right}",
                   " Performing monthly bias correction with the empirical quantile mapping",
                   " method, for each model and month separately. This can take a while. Season",
                   glue::glue_collapse(season, "-")
@@ -189,8 +192,7 @@ projections <-
                                 suppressMessages(
                                   downscaleR::biasCorrection(
                                     y = obs[[1]],
-                                    x = dplyr::filter(datasets, experiment == "historical")$models_mbrs[[1]],
-                                    newdata = mod,
+                                    x = mod,
                                     precipitation = ifelse(var == "pr", TRUE, FALSE),
                                     method = "eqm",
                                     window = if (any(diffs == 1))
@@ -207,100 +209,114 @@ projections <-
                                 mod$Dates$start
                               mod_temp$Dates$end <-  mod$Dates$end
                               if (any(is.na(mod_temp$Data)))
-                                cli::cli_text("{cli::symbol$arrow_right} Bias correction has introduced NA values in certain pixels. Proceed with care")
+                                cli::cli_alert_warning("Bias correction has introduced NA values in certain pixels. Proceed with care")
                               return(mod_temp)
                             }, .progress = T))
             } else
               .
           }  %>%   # computing annual aggregation. if threshold is specified, first apply threshold
-          dplyr::mutate(models_agg_y = furrr::future_map(models_mbrs, function(x)
-            suppressMessages(
-              transformeR::aggregateGrid(# perform aggregation based on seasonended output
-                x, aggr.y =
-                  if (var == "pr" &
-                      !consecutive &
-                      (is.null(uppert) & is.null(lowert))) {
-                    list(FUN = "sum")
-                  } else if (var != "pr" &
-                             !consecutive &
-                             (is.null(lowert) &
-                              is.null(uppert))) {
-                    list(FUN = "mean")
-                  } else if (consecutive) {
-                    list(
-                      FUN = thrs_consec,
-                      duration = duration,
-                      lowert = lowert,
-                      uppert = uppert
-                    )
-                  } else if (!consecutive) {
-                    list(FUN = thrs,
-                         uppert = uppert,
-                         lowert = lowert)
-                  })
-            ))) %>%
-          dplyr::select(-models_mbrs) %>%
-          # ensemble mean
-          dplyr::mutate(
-            rst_ens_mean = purrr::map2(experiment, models_agg_y, function(x, y) {
-              ens <-
-                suppressMessages(transformeR::aggregateGrid(y, aggr.mem = list(FUN = "mean", na.rm = TRUE)))
-              rs <-
-                make_raster(ens, if (length(ens$Dates$start) == 1)
-                  c(1, 2)
-                  else
-                    c(2, 3), country_shp) # adjust by array dimension
-              names(rs) <-
-                paste0(x, "_", names(rs))
-              return(rs)
-            }),
-            # ensemble SD
-            rst_ens_sd = purrr::map2(experiment, models_agg_y, function(x, y) {
-              ens <-
-                suppressMessages(transformeR::aggregateGrid(y, aggr.mem = list(FUN = "sd", na.rm = TRUE)))
-              rs <-
-                make_raster(ens, if (length(ens$Dates$start) == 1)
-                  c(1, 2)
-                  else
-                    c(2, 3), country_shp)
-              names(rs) <-
-                paste0(x, "_", names(rs))
-              return(rs)
-            }),
-            # individual models
-            rst_models = purrr::map2(experiment, models_agg_y, function(x, y) {
-              rs_list <- purrr::map(1:dim(y$Data)[[1]], function(ens) {
-                array_mean <-
-                  if (length(y$Dates$start) == 1)
-                    apply(y$Data[ens, , ,], c(1, 2), mean, na.rm = TRUE)
+          dplyr::mutate_at(c("models_mbrs", "obs"),
+                           ~ furrr::future_map(., ~
+                                                 suppressMessages(
+                                                   transformeR::aggregateGrid(# perform aggregation based on seasonended output
+                                                     ., aggr.y =
+                                                       if (var == "pr" &
+                                                           !consecutive &
+                                                           (is.null(uppert) &
+                                                            is.null(lowert))) {
+                                                         list(FUN = "sum")
+                                                       } else if (var != "pr" &
+                                                                  !consecutive &
+                                                                  (is.null(lowert) &
+                                                                   is.null(uppert))) {
+                                                         list(FUN = "mean")
+                                                       } else if (consecutive) {
+                                                         list(
+                                                           FUN = thrs_consec,
+                                                           duration = duration,
+                                                           lowert = lowert,
+                                                           uppert = uppert
+                                                         )
+                                                       } else if (!consecutive) {
+                                                         list(FUN = thrs,
+                                                              uppert = uppert,
+                                                              lowert = lowert)
+                                                       })
+                                                 ))) %>%
+          dplyr::mutate(rst_obs = purrr::map(obs, function(data_obs) {
+            rs <-
+              make_raster(data_obs, if (length(data_obs$Dates$start) == 1)
+                c(1, 2)
                 else
-                  apply(y$Data[ens, , ,], c(2, 3), mean, na.rm = TRUE) # climatology per member adjusting by array dimension
-                y$Data <- array_mean
-                rs <- make_raster(y, c(1, 2), country_shp)
-                names(rs) <-
-                  paste0("Member ", ens, "_", x, "_", names(rs))
-                return(rs)
-              })
+                  c(2, 3), country_shp)
+          })) %>%
+          # individual models rasters
+          dplyr::mutate(rst_models = purrr::map(models_mbrs, function(y) {
+            rs_list <- purrr::map(1:dim(y$Data)[[1]], function(ens) {
+              array_mean <-
+                if (length(y$Dates$start) == 1)
+                  apply(y$Data[ens, , , ], c(1, 2), mean, na.rm = TRUE)
+              else
+                apply(y$Data[ens, , , ], c(2, 3), mean, na.rm = TRUE) # climatology per member adjusting by array dimension
+
+              y$Data <- array_mean
+
+              rs <- make_raster(y, c(1, 2), country_shp)
+
+              names(rs) <-
+                paste0("Member ", ens, "_historical_", names(rs))
+              return(rs)
+            })
+          })) %>%
+          # biases
+          dplyr::mutate(
+            rst_mod_biases = purrr::map(rst_models, function(y) {
+              biases <- terra::rast(y) - rst_obs[[1]]
+              return(biases)
+            }),
+            rst_ens_biases = purrr::map(rst_mod_biases, function(y) {
+              mean_biases <- terra::mean(y)
+              names(mean_biases) <- names(terra::rast(y))[[1]]
+              return(mean_biases)
+            }),
+            models_temp = purrr::map(models_mbrs, function(x) {
+              dimnames(x$Data)[[1]] <- x$Members
+              dimnames(x$Data)[[2]] <- x$Dates$start
+              dimnames(x$Data)[[3]] <- x$xyCoords$y
+              dimnames(x$Data)[[4]] <- x$xyCoords$x
+
+              obs_data <- apply(obs[[1]]$Data, 1, mean, na.rm = T)
+              obs_date <- as.Date(obs[[1]]$Dates$start)
+
+              df_obs <-
+                data.frame(obs_value = obs_data, date = obs_date)
+
+              df <- suppressMessages(
+                reshape2::melt(x$Data) %>%
+                  dplyr::mutate(date = as.Date(Var2)) %>%
+                  dplyr::mutate(experiment = "historical") %>%
+                  dplyr::group_by(Var1, date) %>%
+                  dplyr::summarise(value = mean(value)) %>%
+                  dplyr::left_join(., df_obs)
+              )
+
+              return(df)
 
             })
           )
-
         invisible(structure(
           list(
-            terra::rast(data_list$rst_ens_mean),
-            terra::rast(data_list$rst_ens_sd),
-            terra::rast(purrr::map(
-              data_list$rst_models, ~ terra::rast(.x)
-            ))
+            data_list$rst_ens_biases[[1]],
+            data_list$rst_mod_biases[[1]],
+            data_list$models_temp[[1]]
           ),
-          class = "CAVAanalytics_projections",
+          class = "CAVAanalytics_model_biases",
           components = list(
-            "SpatRaster for ensemble mean",
-            "SpatRaster for ensemble sd",
-            "SpatRaster for individual members"
+            "SpatRaster for ensemble biases",
+            "SpatRaster for model biases",
+            "data frame for temporal biases"
           )
         ))
-
       }
 
     # beginning of code -------------------------------------------------------
@@ -308,7 +324,6 @@ projections <-
       cli::cli_abort(c("x" = "The input data is not the output of CAVAanalytics load_data"))
     # check input requirements
     check_inputs(data, bias.correction, uppert, lowert, consecutive, duration)
-
     # retrieve information
     mod.numb <- dim(data[[1]]$models_mbrs[[1]]$Data) [1]
     datasets <- data[[1]]
@@ -318,6 +333,7 @@ projections <-
     dates <- as.Date(dates)
     # calculate the differences between consecutive dates to understand temporal resolution and adjust the window argument in bias correction
     diffs <- diff(dates)
+
 
     # create message
     mes <-
@@ -329,7 +345,7 @@ projections <-
     # filter data by season
     datasets <- filter_data_by_season(datasets, season)
     cli::cli_text(paste0("{cli::symbol$arrow_right}",
-      " projections, season ",
+      " Model biases, season ",
       glue::glue_collapse(season, "-"),
       ". ",
       mes