+++ R/utils.R
@@ -23,23 +23,22 @@ common_dates <- function(data) {
 #' Make a spatRaster from a C4R list with dim(Data)=2
 #'
 #' @param cl4.object A C4R list with the Data slot in two dimension
+#' @param dimensions vector specifying which dimensions corresponds to lat and lon. The rest will be averaged
+#' @param shape.file sf object for which to crop and mask the spatRaster
 #' @return spatRaster
 #' @export
 
-make_raster <- function(cl4.object) {
-  if (length(dim(cl4.object$Data)) != 2)
-    stop("Your data needs to be a 2d array, check dimension")
-
+make_raster <- function(cl4.object, dimensions, shape.file) {
   xmin <-
     if (is.null(cl4.object$xyCoords$lon))
       min(cl4.object$xyCoords$x)
   else
-    min(cl4.object$xyCoords$lon[1, ])
+    min(cl4.object$xyCoords$lon[1,])
   xmax <-
     if (is.null(cl4.object$xyCoords$lon))
       max(cl4.object$xyCoords$x)
   else
-    max(cl4.object$xyCoords$lon[1, ])
+    max(cl4.object$xyCoords$lon[1,])
   ymin <-
     if (is.null(cl4.object$xyCoords$lat))
       min(cl4.object$xyCoords$y)
@@ -51,9 +50,15 @@ make_raster <- function(cl4.object) {
   else
     max(cl4.object$xyCoords$lat[, 1])
 
+  array_mean <- apply(cl4.object$Data, dimensions, mean, na.rm = TRUE)
+
+  cl4.object$Data <- array_mean
+
   rasters <- terra::rast(cl4.object$Data,
                          extent = terra::ext(xmin, xmax, ymin, ymax)) %>%
-    terra::flip(., direction = 'vertical')
+    terra::flip(., direction = 'vertical') %>%
+    terra::crop(., shape.file, snap = "out") %>%
+    terra::mask(., shape.file)
 
   nms <-
     paste0(
@@ -218,14 +223,14 @@ ens_trends <- function(c4R) {
       out <- anova(mnlm, p.uni = "adjusted")
 
       sig.models <-
-        names(out$uni.p[2, ][out$uni.p[2, ] < 0.05]) # names of models with significance (p.value < 0.05)
+        names(out$uni.p[2,][out$uni.p[2,] < 0.05]) # names of models with significance (p.value < 0.05)
       colnames(mnlm$coefficients) <- paste0("X", 1:mbrs)
       coef_res <- mnlm$coefficients[2,  sig.models]
       prop_res <-
         if (length(coef_res) == 0)
           999
       else
-        sum(ifelse(coef_res  >= 0, 1,-1)) # number of models, with significance, that shows an increaseor decraese. 999 assign to NA
+        sum(ifelse(coef_res  >= 0, 1, -1)) # number of models, with significance, that shows an increaseor decraese. 999 assign to NA
       cbind(prop_res, out$table[2, 4])
 
     })
@@ -244,14 +249,14 @@ ens_trends <- function(c4R) {
     out <- anova(mnlm, p.uni = "adjusted")
 
     sig.models <-
-      names(out$uni.p[2, ][out$uni.p[2, ] < 0.05]) # names of models with significance (p.value < 0.05)
+      names(out$uni.p[2,][out$uni.p[2,] < 0.05]) # names of models with significance (p.value < 0.05)
     colnames(mnlm$coefficients) <- paste0("X", 1:mbrs)
     coef_res <- mnlm$coefficients[2,  sig.models]
     prop_res <-
       if (length(coef_res) == 0)
         999
     else
-      sum(ifelse(coef_res  >= 0, 1,-1)) # number of models, with significance, that shows an increase or decraese. 999 assign to NA
+      sum(ifelse(coef_res  >= 0, 1, -1)) # number of models, with significance, that shows an increase or decraese. 999 assign to NA
 
     df_tm_series <- reshape2::melt(c4R$Data)  %>%
       dplyr::select(-Var2) %>%
@@ -397,4 +402,3 @@ ridgeline <- function(x,
     ggplot2::ylab(ylab) +
     ggplot2::xlab(xlab)
 }
-