+++ NEWS.md
@@ -9,6 +9,10 @@ output:
     toc_depth: 2
 ---
 
+## CAVAanalytics 1.1.3
+- Added function model_biases to look at the delta between observation and model simulations
+- Customized cli messages and improved clarity
+
 ## CAVAanalytics 1.1.2
 - Changed appearance of temporal trends when spatial.aggr=T
 - Made resulting ggplot object of plotting more easily customizable